*** Added workaround for a bug on Kawa *** See "acvpoeov" (Tue, 23 Jul 2019 11:37:32 +0900)
Added Secretary pattern to manage all thread accessing to the Scheme
object and it happened to be revealed that there is a possible bug ( or
could be a specification ) that every thread must be properly
initialized before accessing to the scheme object; otherwise some
critical behavior will be performed by the scheme object.

The current biggest problem for Pulsar is calling gui-build! causes to
strangely returns multiple values to those functions which called
gui-build function. The reason is unknown for now.

This could be a bug. But it is necessary to investigate further
causalities.

In order to keep the reproducibility, we are going to commit the current
state before we fix the problem.