Partial commit July 24 2019 in the process of integrating the initialization process of Pulsar.
Yesterday I could not correctly imagine what the scheme initialization
on the Pulsar system supposed to be. Yesterday I categorized those many
initializers into two : dynamic/static but the way to categorize did not
seem to be proper. It did not work properly in my mind and I could not
figure out how.

This morning I realized that I am struggling with the objects which have
totally uneven life cycle and each of those objects supposed to have two
kind of initializers.

Please see memorandums which are left on SchemeSecretary.java