Added target-notation? to (n).
- Note that values on the mapvals should not be consumed in case that
  some notes are skipped due to not belonging to the target.

- Now (n) duplicates all values on the list.

- Now (ls) returns list of MetroTrack objects themself, not of their names.

We deprecated (m!). It became a problem because m! mutates the objects
on the arguments, in the meantime (n) does not mutate. And some part of
code expect m! to mutate the value. For example, (m! (filter f list) bra
bra) works while (n (filter f list) bra bra) does not.

This is the reason why we need to modify the specification of (n).