Modified the help system :
The formely it was not able to add new categories to the help system. It
was not scalable. It was because the help system is built upon the
Java's Enum; therefore, the number of elements is statically defined and
cannot be changed.

By this modification, it was released from the Enum's restriction and
users can freely define new categories. And there came a new drawback;
now it is necessary to initialize these classes to invoke the help
initialization methods since they are separately defined now. Otherwise
these classes are not automatically loadded.

Please refer pulsar.PulsarApplication.loadAllAvailableHelps() for
further information.