Modified the behavior of the Extend Selection.
Now the action is able to shrink the current selection when the cursor
key of the opposite direction is pressed.

TODO: this time, I wrote in the routine for the parenthesis-wise
movement, and hoped that the lisp-word-wise movement also works as well.
But it didn't work in that way. Fix it later.