Set direct-meeting = true, then no dead-lock was given.
The exact reason of this is not known; however this implies that the
scheme event queueing is not necessary anymore while formerly setting
true to this gave us a lot of dead-locks.

Removal of the event queue system is a good idea. We will do it in the
following commits.