Now "put-track" "remove-track" return Values.empty.
Returning a list which is created by Java DOES cause a problem in Scheme
code. The list will miraculously be merged with the values which are
returned by following code. The merged mysterious value cannot be
checked its type somehow. The exact reason why this happens is unknown.
Googling this gives me no result.

Further investigation is required to fix the problem. In order to avoid
the problem without the further investigation, we decided to return
Values.empty.