Performed a countermeasure for the issue that Process blocks its referrer threads.
This is a simple countermeasure for two issues which is described below :

1.
The class java.lang.Process causes those threads to block when an
instancese of it is referred by process-* procedures.

2.
It seems that if a class has an abstract method, Kawa cannot call the
method unless its user explicitly specifies the type of the variable
which stores the instance of the class.

The intension of this class is simply wrap the abstract class by a
concrete class in order to avoid the problem.