Now Pulsar can record MIDI input and play it.
This commit also includes a modification to reorganize
MetroBufferedMidiReceiver's interface. All methods should return an
object; they currently return a boolean value which denotes if the
current track should continue to play or not.

Therefore we added "end()" method which can be used instead of returning
boolean values. This ultimately effects to PulsarNoteListParser's
behavior; the methods on the PulsarNoteListParser returns boolean values
as well.