Now NoteListParserElement can return multiple values.
While MetroMidiReceiver/MetroBufferedMidiReceiver ( hereafter we call
them "the receivers" ) can only generate one value per a method calling,
the 'noteOnOff' function requires to return two values at once.

I tried to modify the interface of the receivers to fit them to the
requirement of returning multiple values, but it was not successful.
The receivers should be able to be delegated each other due to the
requirement of converting MIDI data and Scheme music list data; the
several ways returning multiple values such as multiple calling
listeners, returning List<> objects are applied but they do not fit to
the delegation. Listener pattern cannot be multiply delegated and its
reason is still unknown.

Therefore I decided to absorb the requirement at NoteListParserElement
which is working outside the receivers.

Now the conversion system is scalable and it is able to extend the the
system into converting between scheme list, binary MIDI data and so on.