A partial commit for creating integrated thread initializers.
Removing unused codes which were used to work for initializing current
Scheme object. It was unknown actually the "current Scheme" was stored
as a ThreadLocal object. It was also unknown that how to initialize it.
How to initialize it, how to maintain the life cycle were all unknown.

It appeared that it is necessary to initialize those ThreadLocal objects
before executing any routine by new threads.

Therefore these facilities to queue the message, parameters for the
message etc. are all unnecessary. And they are now obstruction to
obtain wider overview for designing the new version of framework.

So I decided to make a partial commit here to remove those codes.

This commit can work (almost) properly.