=== Modified the initialization process of the Documentation Tool ===
1. The initScheme() method should take an Environment object, not a Scheme
   object.
2. The initScheme() based initialization is not effective; this should
   be done by `require` mechanism of KawaScheme.
3. At the point of testing this, the documentation for `pulsar-gui` was
   not output; this is because the document object is not created until
   the pulsar-gui library is loaded by `require` method of KawaScheme.
4. Now every DocumentCategory object holds a initializer which invokes
   initScheme() method. This effectively creates its document object;
   causes the documentation to be output properly.

All above is done. Now we get the output from `pulsar-gui`.

Now the situation is half-`initScheme()` way and half-`require` way;
we are going to go to `require` way but in the middle of the road.
This half-situation causes the problem. In near future, this situation
should be corrected.