made Metro single-threaded
I think I give up the multithreading in the current way. This
single-threaded version is far simpler and runs smoother and faster.

Currently Metro has two threads: one is for buffering MIDI data and the
another is for playing the MIDI data. It appeared that some sort of
manipulating tracks is necessary when playing MIDI data; that is, the
data flow is now going both way. The bidirectional data flow makes
the synchronization procedure very complicated.

I think the process playing MIDI data should be split into two: MIDI
data generation and MIDI data playing. And these could be performed by
two different threads.

In this version is functionally same as the former version but it runs
in single-threaded mode.

Most of code for multithreading is not necessary anymore. They are to
be removed in the next version.