reorganized some of method names in Metro/Pulsar and added thread initialization
At first, this modification was intended to counteract to the issue that
in (n type: 'note note: lambda0 )  the lambda0 is not properly executed.
After that, it appeared that it occurs because of the spec of (n).  That
is, this modification is actually invalid.

Though, the new structure is much tidier than the former version; I
decided not to revert and to leave it as it is.